+++ www/modules/caldav/CaldavModule.php
@@ -2,9 +2,7 @@
 
 namespace GO\Caldav;
 
-use GO\Base\Db\PDO;
 use GO\Caldav\Model\DavEvent;
-use GO\Caldav\Model\DavTask;
 use go\core\http\Exception;
 
 class CaldavModule extends \GO\Base\Module {
@@ -18,23 +16,8 @@ class CaldavModule extends \GO\Base\Module {
 		return true;
 	}
 
-	public static function initListeners() {
-		
-//		if(\GO::modules()->isInstalled('calendar')) {
-//			\GO\Calendar\Model\Event::model()->addListener("save", "GO\Caldav\CaldavModule", "saveEvent");
-//			\GO\Calendar\Model\Event::model()->addListener("delete", "GO\Caldav\CaldavModule", "deleteEvent");
-//		}
-		
-//		if(\GO::modules()->isInstalled('tasks'))	{
-//			\GO\Tasks\Model\Task::model()->addListener("save", "GO\Caldav\CaldavModule", "saveTask");
-//			\GO\Tasks\Model\Task::model()->addListener("delete", "GO\Caldav\CaldavModule", "deleteTask");
-//		}
-			
-		
-	}
 	
 	public static function saveEvent(\GO\Calendar\Model\Event $event, $davEvent, $data = null) {
-//		return;
 		if($event->isException()) {
 			return;
 		}
@@ -53,16 +36,13 @@ class CaldavModule extends \GO\Base\Module {
 			}
 
 		} else  {			
-//			$davEvent->uri = $event->getUri();			
-			$davEvent->data = isset($data) ? $data : self::exportCalendarEvent($event);			
+			$davEvent->data = isset($data) ? $data : self::exportCalendarEvent($event);
 			$davEvent->mtime = $event->mtime;
-//			$davEvent->calendarId = $event->calendar_id;
-			if($davEvent->save()){
+			if($davEvent->save()) {
 				return $davEvent;
 			} else{
 				throw new Exception("Could not save DAV event");
 			}
-
 		}		
 	}